Fix path resolution for symlinks (#23072)
If `/a/link` is a symlink to directory `/b/c` then `/a/link/..` should
be resolved to `/b/c/..` which is `/b`. Currently, we cancel out
`link/..` to get `/a`. The problem is that we apply `PATH_FS.resolve`
and `PATH_FS.normalize` to paths. These functions cancel `..`
incorrectly.

This at least partially handles the situation. `lookupPath` is modified
to avoid calls that call `PATH_FS.normalize()`. We check that `mkdir`,
`open`, `stat`, `truncate`, and `chmod` now work correctly.